- hash_tree num_blocks and num_hashes members should be 64 bits wide

+++ hash-tree.h
@@ -3,8 +3,8 @@
 
 struct hash_tree {
 	struct rb_root	root;
-	unsigned int	num_blocks;
-	unsigned int	num_hashes;
+	uint64_t	num_blocks;
+	uint64_t	num_hashes;
 };
 
 struct dupe_blocks_list {
